requirements.txt
fastapi
pillow
uvicorn
requests
pytesseract
git+https://github.com/trusted-dev/secureocr.git
